Attempt to quiet warning on 32bits
git-svn-id: http://root.cern.ch/svn/roottest/trunk@2971 7c735d0d-823a-0410-acfd-cd159556d4bf

+++ root/treeformula/array/twodimvar.C
@@ -125,5 +125,5 @@ void tscan(const char *input) {
 void ss() {
    A a;
    fprintf(stdout,"A::aa=%ld A::aaa=%ld\n",
-           sizeof(a.aa),sizeof(a.aaa));
+           (long)sizeof(a.aa),(long)sizeof(a.aaa));
 }
